Refactor: move config class location
- global.config -> global.config.audit 처럼 도메인에 따라 아래에 패키지를 다시 만듬
